requirements.txt
pandas==2.2.2
scikit-learn==1.5.0
imbalanced-learn==0.12.3
mlflow==2.19.0
click==8.1.7
python-dotenv==1.0.1